Extract transaction manager module
